Dec 11 version

+++ competitionSetupFiles/d.cpp
@@ -4,17 +4,19 @@ using namespace std;
 using ll = long long;
 // const long long MOD = 1e9 + 7;
 const long long MOD = 998244353;
-const int IINF = 2e9 + 5; const long long LLINF = (long long)(9e18) + 5;
+const int IINF = 2e9 + 5; 
+const long long LLINF = (long long)(9e18) + 5;
 
 using vll = vector<long long>;
 using vvll = vector<vector<long long>>;
 using pll = pair<long long, long long>;
 
 void solve() {
-    ll n;
-    cin >> n;
+    int n; cin >> n;
     vll v(n);
-    for (int i = 0; i < n; i++) {cin >> v[i];}
+    for (int i = 0; i < n; i++) {
+        cin >> v[i];
+    }
 }
 
 //------------------------------------------------------
